Unified the ammonia reactions
